+++ lib/lotus-mailer.rb
@@ -1,0 +1,1 @@
+require 'lotus/mailer'
