+++ lib/lotus/mailer/rendering/template_name.rb
@@ -4,30 +4,45 @@ module Lotus
   module Mailer
     module Rendering
       # @since 0.1.0
+      # @api private
+      #
+      # TODO this is identical to Lotus::View, consider to move into Lotus::Utils
       class TemplateName
+        # @since 0.1.0
+        # @api private
         NAMESPACE_SEPARATOR = '::'.freeze
 
+        # @since 0.1.0
+        # @api private
         def initialize(name, namespace)
           @name = name
           compile!(namespace)
         end
 
+        # @since 0.1.0
+        # @api private
         def to_s
           @name
         end
 
         private
+        # @since 0.1.0
+        # @api private
         def compile!(namespace)
           tokens(namespace) {|token| replace!(token) }
           @name = Utils::String.new(@name).underscore
         end
 
+        # @since 0.1.0
+        # @api private
         def tokens(namespace)
           namespace.to_s.split(NAMESPACE_SEPARATOR).each do |token|
             yield token
           end
         end
 
+        # @since 0.1.0
+        # @api private
         def replace!(token)
           @name.gsub!(%r{\A#{ token }#{ NAMESPACE_SEPARATOR }}, '')
         end
