+++ lib/lotus/mailer/version.rb
@@ -1,5 +1,6 @@
 module Lotus
   module Mailer
-    VERSION = '0.1.0'
+    # @since 0.1.0
+    VERSION = '0.1.0'.freeze
   end
 end
