+++ lotus-mailer.gemspec
@@ -26,7 +26,7 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency 'lotus-utils', '~> 0.5'
   spec.add_dependency 'tilt',        '~> 2.0', '>= 2.0.1'
-  spec.add_dependency 'mail'
+  spec.add_dependency 'mail',        '~> 2.5'
 
   spec.add_development_dependency 'bundler', '~> 1.10'
   spec.add_development_dependency 'rake',    '~> 10.0'
